// schematics/src/store-group/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Creates a new store group.
 */
export interface PwaStoreGroupOptionsSchema {
  path?: string;
  restricted?: boolean;
  tsext?: string;
  project?: string;
  module?: string;
  /**
   * The name of the store group.
   */
  name?: string;
}